Restructure client, fix STAR splitting
