+++ tests/test_tools.py
@@ -1,8 +1,7 @@
 import pytest
 
-import hash
 import os
-import s3
+from tests.util import s3, hash
 
 if os.environ.get("DEPLOY_ENVIRONMENT") == "staging":
     os.environ["BASE_URL"] = os.environ["TOOLCHEST_STAGING_URL"]
