+++ tests/util/hash.py
