+++ toolchest_client/__init__.py
@@ -10,7 +10,7 @@ load_dotenv(find_dotenv(".env"))
 # specifying print flushing is necessary to support loading from R
 builtins.print = functools.partial(print, flush=True)
 
-from .auth import get_key, set_key
-from .exceptions import ToolchestException, DataLimitError, ToolchestJobError
-from .query import Query
+from toolchest_client.api.auth import get_key, set_key
+from toolchest_client.api.exceptions import ToolchestException, DataLimitError, ToolchestJobError
+from toolchest_client.api.query import Query
 from .tools.api import bowtie2, cutadapt, kraken2, STAR, test, unicycler
