+++ toolchest_client/files/tests/data/eight_line_split_one.fastq
@@ -1,0 +1,4 @@
+@test.fastq.1 BOGUS length=121
+ACTGTTCGATGNACTGTTCGATGNACTGTTCGATGNACTGTTCGATGNACTGTTCGATGNACTGTTCGATGNACTGTTCGATGNACTGTTCGATGNACTGTTCGATGNACTGTTCGATGN
++test.fastq.1 BOGUS length=121
+=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<<
