+++ toolchest_client/files/tests/data/eight_line_split_two.fastq
@@ -1,0 +1,4 @@
+@test.fastq.2 BOGUS length=121
+TGTTCGATGNACTGTTCGATTGTTCGATGNACTGTTCGATTGTTCGATGNACTGTTCGATTGTTCGATGNACTGTTCGATTGTTCGATGNACTGTTCGATTGTTCGATGNACTGTTCGAT
++test.fastq.2 BOGUS length=121
+=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<=<<=====<=<==<<
