+++ toolchest_client/tools/kraken2.py
@@ -5,7 +5,7 @@ toolchest_client.tools.kraken2
 This is the Kraken2 implementation of the Tool class.
 """
 from . import Tool
-from ..files import concatenate_files
+from toolchest_client.files import concatenate_files
 
 
 class Kraken2(Tool):
